backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-generativeai==0.3.2
PyMuPDF==1.23.8
python-dotenv==1.0.0
Werkzeug==3.0.1